Refactor project hooks: extract socials, members, and permissions logic
